fix: storage_limit_bytes INTEGER → BIGINT (> 2 GB Werte)
INTEGER (int32) hat max ~2.1 GB — für 5 GB / 25 GB Pläne reicht das nicht.
Alle 4 Storage-Spalten auf BIGINT hochgestuft.
Migration ALTER COLUMN TYPE BIGINT für bestehende INTEGER-Spalten.

+++ backend/models/customer.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime, date
-from sqlalchemy import String, Boolean, DateTime, Date, Numeric, Integer, ForeignKey, Text
+from sqlalchemy import String, Boolean, DateTime, Date, Numeric, Integer, BigInteger, ForeignKey, Text
 from sqlalchemy.dialects.postgresql import UUID, JSONB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from core.database import Base
@@ -20,7 +20,7 @@ class SubscriptionPlan(Base):
     daily_token_limit: Mapped[int | None] = mapped_column(Integer, nullable=True)  # Tokens pro Tag (0 = kein Limit)
     requests_per_hour: Mapped[int | None] = mapped_column(Integer, nullable=True)  # Max Anfragen pro Stunde
     token_overage_chf_per_1k: Mapped[float] = mapped_column(Numeric(8, 4), default=0.002)
-    storage_limit_bytes: Mapped[int] = mapped_column(Integer, default=524_288_000)  # 500 MB default
+    storage_limit_bytes: Mapped[int] = mapped_column(BigInteger, default=524_288_000)  # 500 MB default
     stripe_price_id_monthly: Mapped[str | None] = mapped_column(String(100), nullable=True)
     stripe_price_id_yearly: Mapped[str | None] = mapped_column(String(100), nullable=True)
     sort_order: Mapped[int] = mapped_column(Integer, default=0)
@@ -75,9 +75,9 @@ class Customer(Base):
     memory_consent: Mapped[bool] = mapped_column(Boolean, default=True)
 
     # Speicher
-    storage_used_bytes: Mapped[int] = mapped_column(Integer, default=0)           # aktuell belegter Speicher
-    storage_limit_bytes: Mapped[int] = mapped_column(Integer, default=524_288_000) # Limit (default 500 MB)
-    storage_extra_bytes: Mapped[int] = mapped_column(Integer, default=0)           # zusätzlich gekaufter Speicher
+    storage_used_bytes: Mapped[int] = mapped_column(BigInteger, default=0)           # aktuell belegter Speicher
+    storage_limit_bytes: Mapped[int] = mapped_column(BigInteger, default=524_288_000) # Limit (default 500 MB)
+    storage_extra_bytes: Mapped[int] = mapped_column(BigInteger, default=0)           # zusätzlich gekaufter Speicher
 
     # Wallet — Prepaid-Guthaben (Token-Overage + externe Zahlungen)
     token_balance_chf: Mapped[float] = mapped_column(Numeric(10, 4), default=0.0)   # alias: wallet_balance_chf
